refactor: vehicles inherit from GCVehicleEntity
